test(kfx): consolidate identity-neutral reverse scan
