+++ package.json
@@ -265,8 +265,8 @@
     "test:kfx-profile-suite": "node scripts/require-shifu.mjs test:kfx-profile-suite && node scripts/run-kfx-profile-suite-tests.mjs",
     "test:mission-authority-cutover": "node scripts/require-shifu.mjs test:mission-authority-cutover && node scripts/run-mission-authority-cutover-tests.mjs",
     "test:agent-review-continuation": "node scripts/require-shifu.mjs test:agent-review-continuation && node scripts/run-agent-review-continuation-tests.mjs",
-    "test:native-kfx-admission": "node scripts/require-shifu.mjs test:native-kfx-admission && node scripts/check-kfx-identity-neutral-authority.mjs && node scripts/run-native-kfx-admission-tests.mjs",
-    "check:kfx-identity-neutral-authority": "node scripts/require-shifu.mjs check:kfx-identity-neutral-authority && node scripts/check-kfx-identity-neutral-authority.mjs",
+    "test:native-kfx-admission": "node scripts/require-shifu.mjs test:native-kfx-admission && node scripts/run-native-kfx-admission-tests.mjs",
+    "check:kfx-identity-neutral-authority": "node scripts/require-shifu.mjs check:kfx-identity-neutral-authority && node scripts/run-native-kfx-admission-tests.mjs --identity-neutral-only",
     "test:profile-lifecycle": "node scripts/require-shifu.mjs test:profile-lifecycle && node scripts/run-profile-lifecycle-tests.mjs",
     "test:agent-profile-sdk": "node scripts/require-shifu.mjs test:agent-profile-sdk && node scripts/run-agent-profile-sdk-tests.mjs",
     "test:agent-console-contract": "node scripts/require-shifu.mjs test:agent-console-contract && node scripts/run-agent-console-contract-tests.mjs",
